requirements.txt
google-genai
python-dotenv
feedparser
requests